# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Mini Event Management System API
  description: >
    RESTful API for browsing events, booking tickets, and recording attendance.
    Built with Node.js + Express + MySQL for the Junior Node.js Developer selection test.
  version: 1.0.0
  contact:
    email: hr@thelattice.in

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Events
    description: Event management endpoints
  - name: Bookings
    description: Ticket booking endpoints
  - name: Users
    description: User-specific endpoints
  - name: Attendance
    description: Event attendance check-in

paths:

  # ── Events ────────────────────────────────────────────────────
  /events:
    get:
      tags: [Events]
      summary: List all upcoming events
      description: Returns all events whose date is in the future, ordered by date ascending.
      operationId: listEvents
      responses:
        '200':
          description: A list of upcoming events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
              example:
                success: true
                data:
                  - id: 1
                    title: Tech Summit 2026
                    description: Annual technology conference
                    date: '2026-06-15T09:00:00.000Z'
                    total_capacity: 200
                    remaining_tickets: 198
                    created_at: '2026-04-01T10:00:00.000Z'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Events]
      summary: Create a new event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventBody'
            example:
              title: Node.js Workshop
              description: Hands-on Node.js deep-dive session
              date: '2026-05-10T10:00:00Z'
              capacity: 50
      responses:
        '201':
          description: Event created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
              example:
                success: true
                data:
                  id: 2
                  title: Node.js Workshop
                  description: Hands-on Node.js deep-dive session
                  date: '2026-05-10T10:00:00.000Z'
                  total_capacity: 50
                  remaining_tickets: 50
                  created_at: '2026-04-02T08:30:00.000Z'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Bookings ──────────────────────────────────────────────────
  /bookings:
    post:
      tags: [Bookings]
      summary: Book a ticket for a user
      description: >
        Books one ticket for the specified user and event.
        Uses a database transaction with SELECT FOR UPDATE to prevent race conditions.
        Returns a unique booking_code the user can present at the event.
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingBody'
            example:
              user_id: 1
              event_id: 1
      responses:
        '201':
          description: Booking created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
              example:
                success: true
                data:
                  booking_id: 7
                  user_id: 1
                  event_id: 1
                  booking_code: 550e8400-e29b-41d4-a716-446655440000
                  booking_date: '2026-04-02T09:00:00.000Z'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: No tickets remaining or user already booked this event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: No tickets remaining for this event.
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Users ─────────────────────────────────────────────────────
  /users/{id}/bookings:
    get:
      tags: [Users]
      summary: Get all bookings for a specific user
      operationId: getUserBookings
      parameters:
        - name: id
          in: path
          required: true
          description: The user ID.
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: User and their bookings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBookingsResponse'
              example:
                success: true
                data:
                  user:
                    id: 1
                    name: Alice Johnson
                    email: alice@example.com
                  bookings:
                    - id: 7
                      booking_code: 550e8400-e29b-41d4-a716-446655440000
                      booking_date: '2026-04-02T09:00:00.000Z'
                      event_id: 1
                      title: Tech Summit 2026
                      description: Annual technology conference
                      event_date: '2026-06-15T09:00:00.000Z'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Attendance ────────────────────────────────────────────────
  /events/{id}/attendance:
    post:
      tags: [Attendance]
      summary: Check in a ticket holder at an event
      description: >
        Validates the booking_code for the given event and records an attendance entry.
        Returns the total number of tickets booked for the event.
      operationId: recordAttendance
      parameters:
        - name: id
          in: path
          required: true
          description: The event ID.
          schema:
            type: integer
            minimum: 1
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceBody'
            example:
              booking_code: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '201':
          description: Attendance recorded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceResponse'
              example:
                success: true
                data:
                  event_id: 1
                  event_title: Tech Summit 2026
                  tickets_booked: 45
                  total_capacity: 200
                  entry_time: '2026-06-15T09:15:00.000Z'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Booking already checked in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: This booking has already been checked in.
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

# ── Components ────────────────────────────────────────────────
components:
  schemas:

    # ── Domain models ──────────────────────────────────────────
    Event:
      type: object
      properties:
        id:                { type: integer, example: 1 }
        title:             { type: string,  example: Tech Summit 2026 }
        description:       { type: string,  example: Annual technology conference }
        date:              { type: string, format: date-time, example: '2026-06-15T09:00:00.000Z' }
        total_capacity:    { type: integer, example: 200 }
        remaining_tickets: { type: integer, example: 198 }
        created_at:        { type: string, format: date-time }

    BookingItem:
      type: object
      properties:
        id:           { type: integer }
        booking_code: { type: string, format: uuid }
        booking_date: { type: string, format: date-time }
        event_id:     { type: integer }
        title:        { type: string }
        description:  { type: string }
        event_date:   { type: string, format: date-time }

    # ── Request bodies ─────────────────────────────────────────
    CreateEventBody:
      type: object
      required: [title, date, capacity]
      properties:
        title:
          type: string
          example: Node.js Workshop
        description:
          type: string
          example: Hands-on Node.js deep-dive session
        date:
          type: string
          format: date-time
          example: '2026-05-10T10:00:00Z'
        capacity:
          type: integer
          minimum: 1
          example: 50

    CreateBookingBody:
      type: object
      required: [user_id, event_id]
      properties:
        user_id:  { type: integer, minimum: 1, example: 1 }
        event_id: { type: integer, minimum: 1, example: 1 }

    AttendanceBody:
      type: object
      required: [booking_code]
      properties:
        booking_code:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000

    # ── Response schemas ───────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }

    ValidationErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        errors:
          type: array
          items:
            type: object
            properties:
              msg:   { type: string }
              param: { type: string }
              location: { type: string }

    EventListResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/Event'

    EventResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:    { $ref: '#/components/schemas/Event' }

    BookingResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            booking_id:   { type: integer }
            user_id:      { type: integer }
            event_id:     { type: integer }
            booking_code: { type: string, format: uuid }
            booking_date: { type: string, format: date-time }

    UserBookingsResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            user:
              type: object
              properties:
                id:    { type: integer }
                name:  { type: string }
                email: { type: string, format: email }
            bookings:
              type: array
              items:
                $ref: '#/components/schemas/BookingItem'

    AttendanceResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            event_id:       { type: integer }
            event_title:    { type: string }
            tickets_booked: { type: integer }
            total_capacity: { type: integer }
            entry_time:     { type: string, format: date-time }

  # ── Reusable responses ─────────────────────────────────────
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Event not found.

    ValidationError:
      description: Input validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'

    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Internal Server Error